Allow the script to do the user change.
This is nice, as it means you can open any pid files
with the permission of the user you run logstash as (i.e. root
from an init script), and then change to a lower priv PID
that can't re-write those files, meaning you don't need to make
a directory for the pids that the user running logstash can write
into.. (Or run logstash as root - but that'd be insane).